@@ -25,19 +25,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "15.2.8",
-    "@angular/animations": "15.2.9",
+    "@angular-devkit/build-angular": "15.2.10",
+    "@angular/animations": "15.2.10",
     "@angular/cdk": "15.2.9",
-    "@angular/common": "15.2.9",
-    "@angular/compiler": "15.2.9",
-    "@angular/core": "15.2.9",
-    "@angular/forms": "15.2.9",
+    "@angular/common": "15.2.10",
+    "@angular/compiler": "15.2.10",
+    "@angular/core": "15.2.10",
+    "@angular/forms": "15.2.10",
     "@angular/material": "15.2.9",
-    "@angular/platform-browser": "15.2.9",
-    "@angular/platform-browser-dynamic": "15.2.9",
+    "@angular/platform-browser": "15.2.10",
+    "@angular/platform-browser-dynamic": "15.2.10",
     "@fortawesome/angular-fontawesome": "0.12.1",
     "@fortawesome/fontawesome-common-types": "6.1.0",
     "@fortawesome/fontawesome-svg-core": "6.1.0",
+    "@fortawesome/free-brands-svg-icons": "6.1.0",
     "@fortawesome/free-regular-svg-icons": "6.1.0",
     "@fortawesome/free-solid-svg-icons": "6.1.0",
     "core-js": "3.31.0",
@@ -55,26 +56,26 @@
   },
   "devDependencies": {
     "@angular-eslint/eslint-plugin": "~15.2",
-    "@angular/cli": "15.2.8",
-    "@angular/compiler-cli": "15.2.9",
-    "@angular/language-service": "15.2.9",
+    "@angular/cli": "15.2.10",
+    "@angular/compiler-cli": "15.2.10",
+    "@angular/language-service": "15.2.10",
     "@types/luxon": "~3.3.0",
-    "@types/node": "16.18.36",
+    "@types/node": "16.18.72",
     "@types/xml2js": "^0.4.11",
-    "@typescript-eslint/eslint-plugin": "~5.59.11",
-    "@typescript-eslint/parser": "~5.59.11",
-    "eslint": "~8.42.0",
-    "eslint-plugin-import": "2.27.5",
+    "@typescript-eslint/eslint-plugin": "~5.62.0",
+    "@typescript-eslint/parser": "~5.62.0",
+    "eslint": "~8.56.0",
+    "eslint-plugin-import": "2.29.1",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "eslint-plugin-prettier": "4.2.1",
     "gts": "~4.0.1",
     "husky": "~8.0.3",
-    "lint-staged": "~13.2.2",
+    "lint-staged": "~13.3.0",
     "npm-run-all": "4.1.5",
-    "postcss": "~8.4.24",
+    "postcss": "~8.4.33",
     "postcss-sass": "0.5.0",
-    "stylelint": "~15.7.0",
+    "stylelint": "~15.11.0",
     "stylelint-config-sass-guidelines": "~10.0.0",
     "stylelint-config-standard": "~33.0.0",
     "stylelint-config-standard-scss": "^9.0.0",
